Move main layout to its own component
