+++ src/scenes/Root/Root.tsx
@@ -1,5 +1,4 @@
 import loadable from '@loadable/component';
-import { makeStyles } from '@material-ui/core';
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import { NotFoundRoute } from '../../components/Error';
@@ -7,10 +6,8 @@ import { useIdentifyInLogRocket } from '../../components/Session';
 import { Authentication } from '../Authentication';
 import { Home } from '../Home';
 import { AppMetadata } from './AppMetadata';
-import { CreateDialogProviders } from './Creates';
 import { CssBaseline } from './CssBaseline';
-import { Header } from './Header';
-import { Sidebar } from './Sidebar';
+import { MainLayout } from './MainLayout';
 import { useNonProdWarning } from './useNonProdWarning';
 
 const Partners = loadable(() => import('../Partners'), {
@@ -41,37 +38,25 @@ const SearchResults = loadable(() => import('../SearchResults'), {
   resolveComponent: (m) => m.SearchResults,
 });
 
-const useStyles = makeStyles(() => ({
-  app: {
-    flex: 1,
-    display: 'flex',
-    height: '100vh',
-  },
-  main: {
-    flex: 1,
-    display: 'flex',
-    flexDirection: 'column',
-  },
-}));
-
 export const Root = () => {
-  const classes = useStyles();
   useNonProdWarning();
   useIdentifyInLogRocket();
 
   const routes = (
     <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="/partners/*" element={<Partners />} />
-      <Route path="/projects/*" element={<Projects />} />
-      <Route path="/engagements/*" element={<Engagements />} />
-      <Route path="/products/*" element={<Products />} />
-      <Route path="/progress-reports/*" element={<ProgressReports />} />
-      <Route path="/languages/*" element={<Languages />} />
-      <Route path="/users/*" element={<Users />} />
-      <Route path="/search" element={<SearchResults />} />
-      <Route path="/locations/*" element={<Locations />} />
-      {NotFoundRoute}
+      <Route key="main" element={<MainLayout />}>
+        <Route path="/" element={<Home />} />
+        <Route path="/partners/*" element={<Partners />} />
+        <Route path="/projects/*" element={<Projects />} />
+        <Route path="/engagements/*" element={<Engagements />} />
+        <Route path="/products/*" element={<Products />} />
+        <Route path="/progress-reports/*" element={<ProgressReports />} />
+        <Route path="/languages/*" element={<Languages />} />
+        <Route path="/users/*" element={<Users />} />
+        <Route path="/search" element={<SearchResults />} />
+        <Route path="/locations/*" element={<Locations />} />
+        {NotFoundRoute}
+      </Route>
     </Routes>
   );
 
@@ -79,17 +64,7 @@ export const Root = () => {
     <>
       <CssBaseline />
       <AppMetadata />
-      <Authentication>
-        <div className={classes.app}>
-          <CreateDialogProviders>
-            <Sidebar />
-          </CreateDialogProviders>
-          <div className={classes.main}>
-            <Header />
-            {routes}
-          </div>
-        </div>
-      </Authentication>
+      <Authentication>{routes}</Authentication>
     </>
   );
 };
